Parse get_subjects reponse body

+++ spec/support/zk_setup.rb
@@ -1,26 +1,30 @@
 zk_dir = File.join CELLECT_ROOT, 'tmp/zookeeper'
 
-`rm -rf #{ zk_dir }; mkdir -p #{ zk_dir }`
-
 CELLECT_ZK_CONFIG = "#{ zk_dir }/zoo.cfg"
-File.open(CELLECT_ZK_CONFIG, 'w') do |out|
-  out.puts <<-TEXT
-    tickTime=2000
-    initLimit=10
-    syncLimit=5
-    dataDir=#{ zk_dir }
-    clientPort=21811
-    forceSync=no
-    snapCount=1000000
-  TEXT
-end
 
-if `echo ruok | nc 127.0.0.1 21811`.chomp == 'imok'
-  pid = `ps aux | grep -e 'Cellect[\/]tmp[\/]zookeeper'`.split[1]
-  puts "Killing rogue zookeeper process: #{ pid }..."
-  `kill -s TERM #{ pid }`
-  sleep 1
-end
+unless ENV['ZK_URL']
+  `rm -rf #{ zk_dir }; mkdir -p #{ zk_dir }`
 
-`zkServer start #{ CELLECT_ZK_CONFIG } > /dev/null 2>&1`
-ENV['ZK_URL'] = 'localhost:21811'
+  File.open(CELLECT_ZK_CONFIG, 'w') do |out|
+    out.puts <<-TEXT
+      tickTime=2000
+      initLimit=10
+      syncLimit=5
+      dataDir=#{ zk_dir }
+      clientPort=21811
+      forceSync=no
+      snapCount=1000000
+    TEXT
+  end
+
+
+  if `echo ruok | nc 127.0.0.1 21811`.chomp == 'imok'
+    pid = `ps aux | grep -e 'Cellect[\/]tmp[\/]zookeeper'`.split[1]
+    puts "Killing rogue zookeeper process: #{ pid }..."
+    `kill -s TERM #{ pid }`
+    sleep 1
+  end
+
+  `zkServer start #{ CELLECT_ZK_CONFIG } > /dev/null 2>&1`
+  ENV['ZK_URL'] ||= 'localhost:21811'
+end
